Move files to minecode_pipelines #660
