+++ minecode_pipelines/pipes/java_stream.py
